requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
pillow==11.2.1
setuptools==78.1.0
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0+cpu
torchvision==0.21.0
typing_extensions==4.13.2
Werkzeug==3.1.3
